Pelican transport fixes and generated after configs
Fixed an issue with common transports for some reason not getting correct value from defaults.yaml. Also addressed issue moving the common transports to the config sub-module and allowed initClient and initServer to recreate the common transport object. In addition, moved the TLSSkipVerify to only run when setupTransfer is called again (either first time `GetTransfer()` is called or `initClient()` or `initServer()` is called).
