+++ origin_ui/advertise.go
@@ -26,7 +26,7 @@ import (
 	"net/url"
 	"time"
 
-	"github.com/pelicanplatform/pelican/client"
+	"github.com/pelicanplatform/pelican/config"
 	"github.com/pelicanplatform/pelican/director"
 	"github.com/pelicanplatform/pelican/param"
 	"github.com/pkg/errors"
@@ -122,7 +122,7 @@ func AdvertiseOrigin() error {
 
 	// We should switch this over to use the common transport, but for that to happen
 	// that function needs to be exported from pelican
-	tr := client.GetTransport()
+	tr := config.GetTransport()
 	client := http.Client{Transport: tr}
 
 	resp, err := client.Do(req)
